Refactor

+++ versati/app/src/main/kotlin/com/liuvil/versati/framework/database/constants.kt
@@ -1,5 +1,5 @@
 package com.liuvil.versati.framework.database
 
 object EntityConstants {
-    val KEY_AUTO_GENERATED: Int = 0
+    val KEY_AUTO_GENERATED: Long = 0
 }
